requirements.txt
albumentations==1.3.1
einops==0.7.0
ffmpeg==1.4
matplotlib==3.8.3
opencv-python==4.8.0.76
Pillow==9.0.0
scikit-learn
scipy==1.11.4
tensorboardX==2.6.2.2
torchsummary==1.5.1
torchvision==0.14.1+cu117
tqdm==4.65.2
torch==1.13.1
torchaudio==2.2.2